Modified the MenuService in order to save the parsed items
The endpoint ultimately fails due to the incorrect mapping. Menu model class needs to be adjusted and new tables need to be created. ***I have forgot to commit the changed service into main and push. As a result I have added Dish and SideDish tables as well as modified  the configurations in order to accept the parsed data ***