Modified IsActive property and added IsMenuActiveAsync
This method fetches the menu by menuId. If it exists, it returns the value of the IsActive property, which is true if the current UTC time is within the StartDate and EndDate range, or false otherwise. The IsActive property checks if the current time falls within the menu's active period (StartDate to EndDate)